Created database folder and moved database_setup.py into it.  Updated app.py
